oracle is a real bastard - basically, because SQLT_INT is the only numeric return type and the rest are chars, this is a fast fix...the proper way would be to have it return SQLT_NUM (a char array) and convert accordingly

+++ engine/server/library/serverDatabase/src/shared/CharacterNameLocator.cpp
@@ -39,7 +39,7 @@ bool CharacterNameLocator::locateObjects(DB::Session *session, const std::string
 				break;
 
 			m_characterIds.push_back(i->character_id.getValue());
-			m_stationIds.push_back(i->character_station_id.getValue());
+			m_stationIds.push_back(static_cast<int>(i->character_station_id.getValue()));
 			m_characterNames.push_back(i->character_name.getValueASCII());
 			m_characterFullNames.push_back(i->character_full_name.getValueASCII());
 			m_characterCreateTime.push_back(i->character_create_time.getValue() + (7 * 60 * 60));
